subsys: bluetooth: services: add workaround for throughput drop issue
When the application buffer is full and the controller buffer is empty, the
application sends all the buffer at once to the controller buffer then the
controller will try to send all received HCI packets to the peer in the whole
connection event.
This is to address the controller does not free the TX buffer as soon as possible
in the middle of connection event.

Signed-off-by: Aaron Ye <aye@ambiq.com>