driver: bluetooth: hci_ambiq: Add Apollo3x throughput drop workaround
For some reason the Apollo3x BLE controller does not free the TX buffer
as soon as possible in the middle of connection event, and causes the
throughput drop issue when retransmission happens. The workaround is to
cached the ACL packets at first and send all the cached ACL packets at
once to the controller when the cached buffer is full and the controller
buffer is empty. Then the controller is able to send all received HCI
buffer to the peer in the whole connection event.

Signed-off-by: Aaron Ye <aye@ambiq.com>