scripts: dts: add oneOf feature for property
Allow choose from one of the type listed under a property binding
using the oneof: key. An optional name key could be added to further
distinguish each options but required for options that have the same type.
All property types are validated the same as before under oneof, but are
not permitted on the same level if oneof is present.

Signed-off-by: Swift Tian <swift.tian@ambiq.com>