boards: shields: ap510_disp: Enable LVGL touchscreen input
This commit enables the touchscreen on the ap510_disp shield to be
used as a pointer input device for LVGL.

A `zephyr,lvgl-pointer-input` devicetree node is added to the shield's
overlay file. This node links the `chsc5x` touchscreen controller as
the input source for the LVGL pointer driver.

With this change, touch events from the display are correctly routed
to LVGL, enabling touch-based UI interaction out-of-the-box when
this shield is used.

Signed-off-by: Zhengwei Wang <zwang@ambiq.com>